little updates

+++ audio_recording.m
@@ -10,7 +10,7 @@ prova_vocoder = audiorecorder(Fs,8,1);
 disp('Start speaking.')
 recordblocking(prova_vocoder, duration);
 disp('End of Recording.');
-myspeech = getaudiodata(prova_vocoder, 'int16');
+myspeech = getaudiodata(prova_vocoder, 'double');
 audiowrite('prova_vocoder_16.wav', myspeech, Fs);
